+++ toro-mopub/src/main/java/im/ene/toro/exoplayer/ExoPlayerViewHelper.java
@@ -21,6 +21,7 @@ import android.support.annotation.NonNull;
 import com.google.android.exoplayer2.ExoPlayer;
 import com.google.android.exoplayer2.ui.SimpleExoPlayerView;
 import im.ene.toro.ToroPlayer;
+import im.ene.toro.annotations.RemoveIn;
 import im.ene.toro.helper.ToroPlayerHelper;
 import im.ene.toro.widget.Container;
 
@@ -36,7 +37,7 @@ import static im.ene.toro.exoplayer.ToroExo.with;
 
 public class ExoPlayerViewHelper extends BaseViewHelper<SimpleExoPlayerView> {
 
-  @SuppressWarnings("unused") @Deprecated //
+  @SuppressWarnings("unused") @RemoveIn(version = "3.6.0") @Deprecated  //
   public ExoPlayerViewHelper(Container container, @NonNull ToroPlayer player, @NonNull Uri uri) {
     this(player, uri, null, with(player.getPlayerView().getContext()).getDefaultCreator());
   }