+++ toro-mopub/src/main/java/im/ene/toro/exoplayer/Playable.java
@@ -33,10 +33,11 @@ import com.google.android.exoplayer2.text.TextRenderer;
 import com.google.android.exoplayer2.trackselection.TrackSelectionArray;
 import com.google.android.exoplayer2.ui.SimpleExoPlayerView;
 import im.ene.toro.ToroPlayer;
+import im.ene.toro.annotations.RemoveIn;
 import im.ene.toro.exoplayer.ui.PlayerView;
 import im.ene.toro.media.PlaybackInfo;
 import im.ene.toro.media.VolumeInfo;
-import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
 
 /**
@@ -160,7 +161,8 @@ public interface Playable<T> {
    * @param volume the volume value to be set. Must be a {@code float} of range from 0 to 1.
    * @deprecated use {@link #setVolumeInfo(VolumeInfo)} instead.
    */
-  @Deprecated void setVolume(@FloatRange(from = 0.0, to = 1.0) float volume);
+  @RemoveIn(version = "3.6.0") @Deprecated  //
+  void setVolume(@FloatRange(from = 0.0, to = 1.0) float volume);
 
   /**
    * Obtain current volume value. The returned value is a {@code float} of range from 0 to 1.
@@ -168,7 +170,8 @@ public interface Playable<T> {
    * @return current volume value.
    * @deprecated use {@link #getVolumeInfo()} instead.
    */
-  @Deprecated @FloatRange(from = 0.0, to = 1.0) float getVolume();
+  @RemoveIn(version = "3.6.0") @Deprecated  //
+  @FloatRange(from = 0.0, to = 1.0) float getVolume();
 
   /**
    * Update playback's volume.
@@ -251,7 +254,7 @@ public interface Playable<T> {
   }
 
   /** List of EventListener */
-  class EventListeners extends ArrayList<EventListener> implements EventListener {
+  class EventListeners extends HashSet<EventListener> implements EventListener {
 
     EventListeners() {
     }